// apps/dashboard/src/lib/api/types.ts
export interface Bank {
  id: string
  name: string
  country_code?: string
  bank_url?: string
  description?: string
  country_flag?: string
  open_finance_integrated: boolean
}